A bug fix
* removed potential type mismatches in elastic/Kernels/Local.cpp
* adjusted some variable names according to the naming convention
* copied local and neighbor data from CPU to GPU at the end of initializeCellLocalMatrices (this was the bug)
* added member functions to LTSTree to retrieve total variable and bucket sizes

+++ src/Equations/elastic/Kernels/Local.cpp
@@ -273,7 +273,7 @@ void seissol::kernels::Local::computeBatchedIntegral(ConditionalBatchTableT &tab
     volKrnl.I = const_cast<const real **>((entry.content[*EntityId::Idofs])->getPointers());
 
     unsigned starOffset = 0;
-    for (unsigned i = 0; i < yateto::numFamilyMembers<tensor::star>(); ++i) {
+    for (size_t i = 0; i < yateto::numFamilyMembers<tensor::star>(); ++i) {
       volKrnl.star(i) = const_cast<const real **>((entry.content[*EntityId::Star])->getPointers());
       volKrnl.extraOffset_star(i) = starOffset;
       starOffset += tensor::star::size(i);
